Move gitlab routes to gitlab group
